requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pdfplumber==0.11.0
pandas==2.2.2
openpyxl==3.1.5
python-docx==1.1.2
python-multipart==0.0.9
aiofiles==24.1.0
httpx==0.27.0
pydantic==2.8.0
pydantic-settings==2.4.0
supabase==2.7.0
chardet==5.2.0
jinja2==3.1.4
pillow==10.4.0
python-jose[cryptography]==3.3.0
PyJWT[crypto]==2.12.1
python-magic==0.4.27
odfpy==1.4.1
xlrd==2.0.1
